# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Fast Note Sync Service HTTP API.",
        "title": "Fast Note Sync Service HTTP API",
        "contact": {
            "name": "Haierkeys",
            "url": "https://github.com/haierkeys",
            "email": "haierkeys@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/",
    "paths": {
...
        "/api/vault": {
            "get": {
                "security": [
                    {
                        "UserAuthToken": []
                    }
                ],
                "description": "Get all note vaults for current user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vault"
                ],
                "summary": "Get vault list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Res"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.VaultDTO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "UserAuthToken": []
                    }
                ],
                "description": "Be used to create a new vault or update an existing vault configuration based on the ID in the request parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vault"
                ],
                "summary": "Create or update vault",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Vault Parameters",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.VaultPostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Res"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.VaultDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "UserAuthToken": []
                    }
                ],
                "description": "Permanently delete a specific note vault and all associated notes and attachments",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vault"
                ],
                "summary": "Delete vault",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/app.Res"
                        }
                    }
                }
            }
        },
        "/api/vault/get": {
            "get": {
                "security": [
                    {
                        "UserAuthToken": []
                    }
                ],
                "description": "Get specific vault configuration details by vault ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vault"
                ],
                "summary": "Get vault details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Auth Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Vault ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Res"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.VaultDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "app.Res": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "context": {},
                "data": {},
                "details": {},
                "message": {},
                "status": {
                    "type": "boolean"
                },
                "vault": {}
            }
        },
...
        "dto.VaultDTO": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "fileCount": {
                    "type": "integer"
                },
                "fileSize": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "noteCount": {
                    "type": "integer"
                },
                "noteSize": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "vault": {
                    "type": "string"
                }
            }
        },
        "dto.VaultPostRequest": {
            "type": "object",
            "required": [
                "vault"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "vault": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ShareAuthToken": {
            "type": "apiKey",
            "name": "Share-Token",
            "in": "header"
        },
        "UserAuthToken": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}